string within string task

+++ hacker-rank-array-sum/cmd/array-sum/main.go
@@ -3,14 +3,14 @@ package main
 import "fmt"
 
 func main() {
-	res := arraySum([]int32{1, 2, 3, 4, 5})
+	res := arraySum([]int64{1, 2, 3, 4, 5})
 	fmt.Println(res)
 }
 
-func arraySum(input []int32) int32 {
-	var sum int32
+func arraySum(input []int64) int64 {
+	var sum int64
 	for _, v := range input {
 		sum += v
 	}
-	return sum
+	return int64(sum)
 }
